I'm still on my journey for ability system but now skill enter the game
as the title imply, i added a "skill" system that allow to use an ability if its condition are met. It was a way to separate ability and condition logic. I'm still pondering wether i should do skill chains or how to deal with skill consequences hmmmm i'll think about it